Patching more syscall holes
