+++ src/wasm/syscalls_unsupported.cpp
@@ -9,11 +9,6 @@
 
 namespace wasm {
 
-    I32 s__fork() {
-        util::getLogger()->debug("S - fork");
-        throwException(Runtime::ExceptionTypes::calledUnimplementedIntrinsic);
-    }
-
     I32 s__futex(I32 uaddrPtr, I32 futex_op, I32 val, I32 timeoutPtr, I32 uaddr2Ptr, I32 other) {
         util::getLogger()->debug("S - futex - {} {} {} {} {} {}", uaddrPtr, futex_op, val, timeoutPtr, uaddr2Ptr, other);
         return 0;
@@ -23,100 +18,37 @@ namespace wasm {
     // Signals (ignored)
     // ------------------------
 
-    DEFINE_INTRINSIC_FUNCTION(env, "sigaction", I32, sigaction, I32 a, I32 b, I32 c) {
+    I32 s__sigaltstack(I32 ssPtr, I32 oldSsPtr) {
+        util::getLogger()->debug("S - sigaltstack - {} {}", ssPtr, oldSsPtr);
+        return 0;
+    }
+
+    I32 s__sigaction(I32 a, I32 b, I32 c) {
         util::getLogger()->debug("S - sigaction - {} {} {}", a, b, c);
 
         return 0;
     }
 
-    DEFINE_INTRINSIC_FUNCTION(env, "sigemptyset", I32, sigemptyset, I32 a) {
+    I32 s__sigemptyset(I32 a) {
         util::getLogger()->debug("S - sigemptyset - {}", a);
 
         return 0;
     }
 
-    DEFINE_INTRINSIC_FUNCTION(env, "signal", I32, signal, I32 a, I32 b) {
+    I32 s__signal(I32 a, I32 b) {
         util::getLogger()->debug("S - signal - {} {}", a, b);
 
         return 0;
     }
 
-    DEFINE_INTRINSIC_FUNCTION(env, "siginterrupt", I32, siginterrupt, I32 a, I32 b) {
+    I32 s__siginterrupt(I32 a, I32 b) {
         util::getLogger()->debug("S - siginterrupt - {} {}", a, b);
 
         return 0;
     }
 
-
-    // ---------------------------------------
-    // pthreads - ignored
-    // ---------------------------------------
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_cond_destroy", I32, pthread_cond_destroy, I32 a) {
-        util::getLogger()->debug("S - pthread_cond_destroy - {}", a);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_cond_init", I32, pthread_cond_init, I32 a, I32 b) {
-        util::getLogger()->debug("S - pthread_cond_init - {} {}", a, b);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_cond_signal", I32, pthread_cond_signal, I32 a) {
-        util::getLogger()->debug("S - pthread_cond_signal - {}", a);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_getspecific", I32, pthread_getspecific, I32 a) {
-        util::getLogger()->debug("S - pthread_getspecific - {}", a);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_key_create", I32, pthread_key_create, I32 keyPtr, I32 destructor) {
-        util::getLogger()->debug("S - pthread_key_create - {} {}", keyPtr, destructor);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_key_delete", I32, pthread_key_delete, I32 a) {
-        util::getLogger()->debug("S - pthread_key_delete - {}", a);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_mutex_destroy", I32, pthread_mutex_destroy, I32 a) {
-        util::getLogger()->debug("S - pthread_mutex_destroy - {}", a);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_mutex_init", I32, pthread_mutex_init, I32 a, I32 b) {
-        util::getLogger()->debug("S - pthread_mutex_init - {} {}", a, b);
-
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_mutex_lock", I32, pthread_mutex_lock, I32 a) {
-        util::getLogger()->debug("S - pthread_mutex_lock - {}", a);
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_mutex_unlock", I32, pthread_mutex_unlock, I32 a) {
-        util::getLogger()->debug("S - pthread_mutex_unlock - {}", a);
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_mutex_trylock", I32, emscripten_pthread_mutex_trylock, I32 a) {
-        util::getLogger()->debug("S - pthread_mutex_trylock - {}", a);
-        return 0;
-    }
-
-    DEFINE_INTRINSIC_FUNCTION(env, "pthread_setspecific", I32, pthread_setspecific, I32 a, I32 b) {
-        util::getLogger()->debug("S - pthread_setspecific - {} {}", a, b);
+    I32 s__rt_sigprocmask(I32 how, I32 sigSetPtr, I32 oldSetPtr, I32 sigsetsize) {
+        util::getLogger()->debug("S - rt_sigprocmask - {} {} {} {}", how, sigSetPtr, oldSetPtr, sigsetsize);
 
         return 0;
     }
